separate up040 and (new) up046 with shared super module
