+++ crates/ruff_linter/src/rules/pyupgrade/rules/mod.rs
@@ -14,6 +14,7 @@ pub(crate) use native_literals::*;
 pub(crate) use open_alias::*;
 pub(crate) use os_error_alias::*;
 pub(crate) use outdated_version_block::*;
+pub(crate) use pep695::*;
 pub(crate) use printf_string_formatting::*;
 pub(crate) use quoted_annotation::*;
 pub(crate) use redundant_open_modes::*;
@@ -36,7 +37,6 @@ pub(crate) use use_pep585_annotation::*;
 pub(crate) use use_pep604_annotation::*;
 pub(crate) use use_pep604_isinstance::*;
 pub(crate) use use_pep646_unpack::*;
-pub(crate) use use_pep695_type_alias::*;
 pub(crate) use useless_metaclass_type::*;
 pub(crate) use useless_object_inheritance::*;
 pub(crate) use yield_in_for_loop::*;
@@ -57,6 +57,7 @@ mod native_literals;
 mod open_alias;
 mod os_error_alias;
 mod outdated_version_block;
+mod pep695;
 mod printf_string_formatting;
 mod quoted_annotation;
 mod redundant_open_modes;
@@ -79,7 +80,6 @@ mod use_pep585_annotation;
 mod use_pep604_annotation;
 mod use_pep604_isinstance;
 mod use_pep646_unpack;
-mod use_pep695_type_alias;
 mod useless_metaclass_type;
 mod useless_object_inheritance;
 mod yield_in_for_loop;
